// src/user_interface/internals/enums/user_interface_calculator_characters.ts
/* eslint-disable no-shadow */
/* eslint-disable no-unused-vars */

enum UserInterfaceCalculatorCharacters {
  DIVISION = "/",
  MULTIPLICATION = "*",
  SEVEN = "7",
  EIGHT = "8",
  NINE = "9",
  ADDITION = "+",
  FOUR = "4",
  FIVE = "5",
  SIX = "6",
  SUBTRACTION = "-",
  ONE = "1",
  TWO = "2",
  THREE = "3",
  ZERO = "0",
  POINT = ".",
  CLEAN = "C",
  BACKSPACE = "<",
  EVALUATE = "=",
}

export default UserInterfaceCalculatorCharacters;